Move compute_bought_with_relations management command
Move compute_bought_with_relations command and related util from
front to core. In management command add handled product_ids into
set, instead of list.

Refs SHOOP-2326
