+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Add management command to generate bought with relations
 - Add ``merchant_notes`` text-field to ``Contact``
 - Add ``first_name`` and ``last_name`` fields to ``PersonContact``
 - Add bought with relation to ``ProductCrossSellType``
@@ -58,7 +59,6 @@ Admin
 Front
 ~~~~~
 
-- Add management command to generate bought with relations
 - Fix bug at ``get_visible_products`` filter when orderable_only is False
 - Set template price display options from the customer
 - Fix bug: BasketStorage.finalize() never called delete() correctly
