+++ shoop/core/utils/product_bought_with_relations.py
